Move kubelet config API to e2e_node
This API is only used by e2e_node, moved it out of e2e framework.
